Refactor screen classes
Changes:

- All shared screen and stage classes are moved to the root package
- GameplayScreen is moved to game package
- LeaderBoardScreen is moved to leaderboard package
- JoinLobbyScreen and LobbyScreen are moved to lobby package
- LoginMenuScreen is moved to user package
- LoadingScreen is moved to root package
- TutorialScreen is moved to tutorial package
- MenuScreen and MenuPresenter are moved to menu package

The screen package is now removed
