+++ core/src/main/kotlin/no/ntnu/beardblaster/BaseScreen.kt
@@ -1,4 +1,4 @@
-package no.ntnu.beardblaster.screen
+package no.ntnu.beardblaster
 
 import com.badlogic.gdx.Gdx
 import com.badlogic.gdx.graphics.GL20
@@ -7,7 +7,6 @@ import com.badlogic.gdx.graphics.g2d.SpriteBatch
 import com.badlogic.gdx.scenes.scene2d.Stage
 import ktx.app.KtxScreen
 import ktx.assets.async.AssetStorage
-import no.ntnu.beardblaster.BeardBlasterGame
 
 abstract class BaseScreen(
     val game: BeardBlasterGame,
