+++ core/src/main/kotlin/no/ntnu/beardblaster/game/GameInstance.kt
@@ -17,7 +17,6 @@ import no.ntnu.beardblaster.commons.game.Loot
 import no.ntnu.beardblaster.commons.spell.SpellAction
 import no.ntnu.beardblaster.commons.wizard.Wizard
 import no.ntnu.beardblaster.leaderboard.LeaderBoardRepository
-import no.ntnu.beardblaster.screen.Phase
 import no.ntnu.beardblaster.spell.*
 import no.ntnu.beardblaster.wizard.WizardState
 import java.util.*
