+++ core/src/main/kotlin/no/ntnu/beardblaster/lobby/LobbyScreen.kt
@@ -1,4 +1,4 @@
-package no.ntnu.beardblaster.screen
+package no.ntnu.beardblaster.lobby
 
 import com.badlogic.gdx.graphics.OrthographicCamera
 import com.badlogic.gdx.graphics.g2d.SpriteBatch
@@ -17,15 +17,16 @@ import ktx.log.debug
 import ktx.log.error
 import ktx.log.logger
 import ktx.scene2d.*
+import no.ntnu.beardblaster.BaseScreen
 import no.ntnu.beardblaster.BeardBlasterGame
 import no.ntnu.beardblaster.WORLD_WIDTH
 import no.ntnu.beardblaster.assets.Nls
 import no.ntnu.beardblaster.commons.State
 import no.ntnu.beardblaster.commons.game.Game
 import no.ntnu.beardblaster.game.GameData
+import no.ntnu.beardblaster.game.GameplayScreen
 import no.ntnu.beardblaster.leaderboard.BeardScale
-import no.ntnu.beardblaster.lobby.LobbyHandler
-import no.ntnu.beardblaster.lobby.LobbyRepository
+import no.ntnu.beardblaster.menu.MenuScreen
 import no.ntnu.beardblaster.ui.*
 import java.util.*
