+++ core/src/main/kotlin/no/ntnu/beardblaster/tutorial/TutorialScreen.kt
@@ -1,4 +1,4 @@
-package no.ntnu.beardblaster.screen
+package no.ntnu.beardblaster.tutorial
 
 import com.badlogic.gdx.Gdx
 import com.badlogic.gdx.graphics.OrthographicCamera
@@ -9,8 +9,10 @@ import ktx.assets.async.AssetStorage
 import ktx.log.logger
 import ktx.scene2d.scene2d
 import ktx.scene2d.textButton
+import no.ntnu.beardblaster.BaseScreen
 import no.ntnu.beardblaster.BeardBlasterGame
 import no.ntnu.beardblaster.assets.Nls
+import no.ntnu.beardblaster.menu.MenuScreen
 import no.ntnu.beardblaster.ui.Image
 import no.ntnu.beardblaster.ui.fullSizeTable
 import no.ntnu.beardblaster.ui.get
